Organise the modules inside `.prelude._internal`
